web-api/requirements.txt
arrow==0.10.0
django==1.11
djangorestframework==3.6.2
mysqlclient==1.3.10
